Extract Pipeline interface
