+++ packages/core/src/index.ts
@@ -6,5 +6,5 @@ export * from './engine/index.js'
 export * from './factories.js'
 export * from './types.js'
 export * from './util/index.js'
-export * from './Pipeline.js'
+export * from './DefaultPipeline.js'
 export * from './DefaultTableStore.js'
