+++ packages/core/src/types/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './Pipeline.js'
+export * from './TableStore.js'
